Mise en place des widgets de création de notebook et de treestore

+++ Perl/Sources/administration/widgets/ListeOptionnels.pm
@@ -37,16 +37,173 @@ use constant COL_DRAGABLE      => 8;
 use constant COL_ID            => 9;
 use constant COL_EST_TOP      => 10;
 use constant COL_CONFIG_BASE => 11;
+# Profils associés
+use constant COL_PROFILS       => 12;
 
 #A accorder avec ci-dessus (dans le même ordre)
 use constant LISTE_COL_TYPES  => ('Glib::String', 'Glib::Int', 'Glib::Int',
     	'Glib::Int', 'Glib::String', 'Glib::Boolean', 'Glib::Int', 'Glib::Boolean',
-	'Glib::Boolean', 'Glib::Int', 'Glib::Boolean', 'Glib::Boolean');
+	'Glib::Boolean', 'Glib::Int', 'Glib::Boolean', 'Glib::Boolean', 'Glib::String');
 
 use constant GRAS_NORMAL => 400;
 use constant GRAS_MAIN   => 800;
 
 use constant INVALID_ID => -1;
+
+sub profils_to_assoc {
+	my $profils = shift;
+	my %assoc_profils = ();
+
+	foreach my $profil (@$profils) {
+		foreach my $id ($profil->get_ids) {
+			my $sNom = $profil->get_nom;
+
+			if (defined($assoc_profils{$id})) {
+				$assoc_profils{$id} .= ", $sNom";
+			}
+			else {
+				$assoc_profils{$id} = $sNom;
+			}
+		}
+	}
+
+	return \%assoc_profils;
+}
+
+sub categories_to_treestores {
+	my ($categories, $profils) = @_;
+	my $assoc_profils = profils_to_assoc($profils);
+	return [ map { items_to_treestore($_->get_items, $assoc_profils) } @$categories ];
+}
+
+sub items_to_treestore {
+	my $items = shift;
+	# id_item => profil1, profil2, ..
+	my $assoc_profils = shift;
+
+	my $treestore = Gtk2::TreeStore->new(LISTE_COL_TYPES);
+
+	foreach my $mainitem (@$items) {
+		my $mainiter = $treestore->append(undef);
+#		print $mainitem->get_nom."\n";
+		$treestore->set($mainiter, COL_NOM,   $mainitem->get_nom,
+				COL_MASSE, $mainitem->get_masse,
+				COL_BRAS,  $mainitem->get_bras,
+				COL_BRAS_L,$mainitem->get_bras_l,
+				COL_IMG,   $mainitem->get_img,
+				COL_PRESENT_PESEE, $mainitem->is_present_pesee,
+				COL_CONFIG_BASE, $mainitem->is_config_base,
+				COL_DRAGABLE, $mainitem->is_dragable,
+				COL_ID, $mainitem->get_id,
+				COL_PROFILS, $assoc_profils->{$mainitem->get_id},
+				COL_EST_TOP, Glib::TRUE);
+		my $sousitems = $mainitem->get_items;
+		if (scalar(@$sousitems)) {
+			$treestore->set($mainiter, COL_EST_ITEM, Glib::FALSE, COL_GRAS, GRAS_MAIN);
+			foreach my $item (@$sousitems) {
+				my $iter = $treestore->append($mainiter);
+				$treestore->set($iter, COL_NOM,   $item->get_nom,
+				COL_MASSE, $item->get_masse,
+				COL_BRAS,  $item->get_bras,
+				COL_BRAS_L,$item->get_bras_l,
+				COL_IMG,   $item->get_img,
+				COL_EST_ITEM, Glib::TRUE,
+				COL_GRAS,   GRAS_NORMAL,
+				COL_PRESENT_PESEE, $item->is_present_pesee,
+				COL_CONFIG_BASE, $item->is_config_base,
+				COL_DRAGABLE, $item->is_dragable,
+				COL_ID, INVALID_ID,
+				COL_EST_TOP, Glib::FALSE);
+			}
+		}
+		else {
+			$treestore->set($mainiter, COL_EST_ITEM, Glib::TRUE, COL_GRAS, GRAS_NORMAL);
+		}
+
+	}
+
+	return $treestore;
+}
+
+sub treestore_to_items {
+	my $treestore = shift;
+
+	my @items = ();
+
+	for (my $mainiter = $treestore->get_iter_first; $mainiter; $mainiter = $treestore->iter_next($mainiter)) {
+		my $mainitem = models::MainItem->new(
+			$treestore->get($mainiter, COL_BRAS, COL_MASSE, COL_NOM, COL_BRAS_L,
+				COL_IMG, COL_PRESENT_PESEE, COL_DRAGABLE, COL_CONFIG_BASE));
+
+		my $id = $treestore->get($mainiter, COL_ID);
+#		$base_items[$id] = $mainitem unless ($id == INVALID_ID);
+
+		push @items, $mainitem;
+		for (my $iter = $treestore->iter_children($mainiter); $iter; $iter = $treestore->iter_next($iter)) {
+			$mainitem->add_item(models::Item->new(
+				$treestore->get($iter, COL_BRAS, COL_MASSE, COL_NOM, COL_BRAS_L,
+					COL_IMG, COL_PRESENT_PESEE, COL_DRAGABLE, COL_CONFIG_BASE)));
+		}
+	}
+
+
+	return (\@items);
+}
+
+sub treeview_centrage {
+	return treeview(COL_NOM, COL_MASSE, COL_BRAS, COL_IMG);
+}
+
+sub treeview_present_pesee {
+	return treeview(COL_NOM, COL_MASSE, COL_PRESENT_PESEE);
+}
+
+sub treeview_configbase {
+	return treeview(COL_NOM, COL_MASSE, COL_CONFIG_BASE, COL_DRAGABLE);
+}
+
+sub treeview {
+	my @listecols = @_;
+
+	######################################################
+	#
+	# Construction des onglets de matos
+	#
+	######################################################
+	my $treeview  = Gtk2::TreeView->new;
+	map {$treeview->append_column(_create_column($treeview, $_)); } @listecols;
+
+	my $entry = ['ligne', 'GTK_TARGET_SAME_WIDGET', 0];
+	$treeview->enable_model_drag_source( 'button1-mask', ['move', 'default'], $entry);
+	$treeview->enable_model_drag_dest(['default'],  $entry );
+
+	#cet iter est celui qui est couramment dragué. On l'initialise avec une valeur au pif
+	#pour pouvoir faire $drag_iter->set($niter) dans _drag_get
+	my $drag_iter; #= $treestore->get_iter_first;
+
+	$treeview->signal_connect("drag_data_get", \&_drag_get, \$drag_iter);
+#	$treeview->signal_connect("drag_data_received", \&_drag_received, [$drag_iter, $this->{_BASE_ITEMS}]);
+	$treeview->signal_connect("drag_data_received", \&_drag_received, \$drag_iter);
+	$treeview->signal_connect("drag_data_delete", \&_drag_delete, \$drag_iter);
+	return $treeview;
+	#Qu'on puisse supprimer quoi. Une ligne avec la touche suppr quoi.
+#	$treeview->signal_connect("key_press_event", \&_key_press, $this->{_BASE_ITEMS});
+	$treeview->signal_connect("key_press_event", \&_key_press);
+}
+
+sub scrolled_window {
+	my $treeview = shift;
+	$treeview->set_rules_hint(Glib::TRUE);
+	# Pour mettre des barres défilantes
+	my $scrolledwindow = Gtk2::ScrolledWindow->new;
+	$scrolledwindow->add($treeview);
+	# On affiche les barres de défilement si nécessaire
+	$scrolledwindow->set_policy ('automatic', 'automatic');
+
+	return $scrolledwindow
+}
+
+
 ######## FONCTIONS DEPRECACTED
 
 ##########################
@@ -79,173 +236,20 @@ sub ajoute_item {
 }
 
 
-sub new {
-	my ($class, $categories, $profils, $config_base) = @_;
-	my $this = {_PROFILS => $profils};
-
-	# id => [profils associés]
-	my @base_items = ();
-	my @treestores = ();
-
-	######################################################
-	#
-	# Construction des onglets de matos
-	#
-	######################################################
-	foreach my $categorie (@$categories) {
-		my $items = $categorie->get_items;
-		map { $base_items[$_->get_id] = [] } @$items;
-
-		push @treestores, { treestore => _create_treestore($items),
-			nom => $categorie->get_nom };
-
-	}
-
-	######################################################
-	#
-	# Construction des relations avec les profils
-	#
-	######################################################
-	foreach my $profil (@$profils) {
-		foreach my $id (@{$profil->get_ids}) {
-			push @{$base_items[$id]}, $profil;
-		}
-	}
-
-	$this->{_BASE_ITEMS} = \@base_items;
-	$this->{_TREESTORES} = \@treestores;
-	return bless($this, $class);
-
-}
-
-sub get_listecentrage {
-	my ($this, $notebook) = @_;
-	return $this->get_liste($notebook, COL_NOM, COL_MASSE, COL_BRAS, COL_IMG);
-}
-sub get_listepresentpesee {
-	my ($this, $notebook) = @_;
-	return $this->get_liste($notebook, COL_NOM, COL_MASSE, COL_PRESENT_PESEE);
-}
-sub get_listeconfigbase {
-	my ($this, $notebook) = @_;
-	return $this->get_liste($notebook, COL_NOM, COL_MASSE, COL_CONFIG_BASE, COL_DRAGABLE);
-}
-
-sub get_liste {
-	my $this = shift;
-	my $notebook = shift;
-
-	if ($notebook) {
-		# Suppression des onglets déjà présents s'il y en a
-		my $n = $notebook->get_n_pages;
-		for (my $j = 0; $j < $n; $j++) {
-			$notebook->remove_page(-1);
-		}
-	}
-	else {
-	 	$notebook = Gtk2::Notebook->new;
-	}
-	
-	my @listecols = @_;
-
-	######################################################
-	#
-	# Construction des onglets de matos
-	#
-	######################################################
-	foreach my $cat (@{$this->{_TREESTORES}}) {
-		my $treestore = $cat->{treestore};
-		my $treeview  = Gtk2::TreeView->new_with_model($treestore);
-		map {$treeview->append_column(_create_column($treestore, $_)); } @listecols;
-
-		my $entry = ['ligne', 'GTK_TARGET_SAME_WIDGET', 0];
-		$treeview->enable_model_drag_source( 'button1-mask', ['move', 'default'], $entry);
-		$treeview->enable_model_drag_dest(['default'],  $entry );
-
-		#cet iter est celui qui est couramment dragué. On l'initialise avec une valeur au pif
-		#pour pouvoir faire $drag_iter->set($niter) dans _drag_get
-		my $drag_iter = $treestore->get_iter_first;
-
-		$treeview->signal_connect("drag_data_get", \&_drag_get, $drag_iter);
-		$treeview->signal_connect("drag_data_received", \&_drag_received, [$drag_iter, $this->{_BASE_ITEMS}]);
-		$treeview->signal_connect("drag_data_delete", \&_drag_delete, $drag_iter);
-		#Qu'on puisse supprimer quoi. Une ligne avec la touche suppr quoi.
-		$treeview->signal_connect("key_press_event", \&_key_press, $this->{_BASE_ITEMS});
-
-		# Pour mettre des barres défilantes
-		my $scrolledwindow = Gtk2::ScrolledWindow->new;
-		$scrolledwindow->add($treeview);
-		# On affiche les barres de défilement si nécessaire
-		$scrolledwindow->set_policy ('automatic', 'automatic');
-
-
-		$notebook->append_page ($scrolledwindow, $cat->{nom});
-
-		$scrolledwindow->show_all();
-		$treeview->set_rules_hint(Glib::TRUE);
-	}
-	return $notebook;
-}
-
 sub _create_column {
-	my $treestore = shift;
+	my $treeview = shift;
 	my $nCol = shift;
-	return _new_col_editable($treestore, 'Nom', COL_NOM)                              if $nCol == COL_NOM;
-	return _new_col($treestore, 'Masse', COL_MASSE, COL_EST_ITEM)                     if $nCol == COL_MASSE;
-	return _new_col($treestore, 'Bras', COL_BRAS, COL_EST_ITEM)                       if $nCol == COL_BRAS;
-	return _new_col($treestore, 'Bras latéral', COL_BRAS_L, COL_EST_ITEM)             if $nCol == COL_BRAS_L;
-	return _new_col_editable($treestore, 'Image', COL_IMG)                            if $nCol == COL_IMG;
-	return _new_col_toggle_may_visible($treestore,
+	return _new_col_editable($treeview, 'Nom', COL_NOM)                              if $nCol == COL_NOM;
+	return _new_col($treeview, 'Masse', COL_MASSE, COL_EST_ITEM)                     if $nCol == COL_MASSE;
+	return _new_col($treeview, 'Bras', COL_BRAS, COL_EST_ITEM)                       if $nCol == COL_BRAS;
+	return _new_col($treeview, 'Bras latéral', COL_BRAS_L, COL_EST_ITEM)             if $nCol == COL_BRAS_L;
+	return _new_col_editable($treeview, 'Image', COL_IMG)                            if $nCol == COL_IMG;
+	return _new_col_toggle_may_visible($treeview,
 		'Présent en pesée', COL_PRESENT_PESEE, COL_EST_ITEM)                      if $nCol == COL_PRESENT_PESEE;
-	return _new_col_toggle_may_visible($treestore,
+	return _new_col_toggle_may_visible($treeview,
 		'Déplaçable', COL_DRAGABLE, COL_EST_TOP)                                  if $nCol == COL_DRAGABLE;
-	return _new_col_toggle_may_visible($treestore,
+	return _new_col_toggle_may_visible($treeview,
 		'Présent dans la configuration de base', COL_CONFIG_BASE, COL_EST_TOP)    if $nCol == COL_CONFIG_BASE;
-}
-
-sub _create_treestore {
-	my $items = shift;
-
-	my $treestore = Gtk2::TreeStore->new(LISTE_COL_TYPES);
-
-	foreach my $mainitem (@$items) {
-		my $mainiter = $treestore->append(undef);
-		$treestore->set($mainiter, COL_NOM,   $mainitem->get_nom,
-				COL_MASSE, $mainitem->get_masse,
-				COL_BRAS,  $mainitem->get_bras,
-				COL_BRAS_L,$mainitem->get_bras_l,
-				COL_IMG,   $mainitem->get_img,
-				COL_PRESENT_PESEE, $mainitem->is_present_pesee,
-				COL_CONFIG_BASE, $mainitem->is_config_base,
-				COL_DRAGABLE, $mainitem->is_dragable,
-				COL_ID, $mainitem->get_id,
-				COL_EST_TOP, Glib::TRUE);
-		my $sousitems = $mainitem->get_items;
-		if (scalar(@$sousitems)) {
-			$treestore->set($mainiter, COL_EST_ITEM, Glib::FALSE, COL_GRAS, GRAS_MAIN);
-			foreach my $item (@$sousitems) {
-				my $iter = $treestore->append($mainiter);
-				$treestore->set($iter, COL_NOM,   $item->get_nom,
-				COL_MASSE, $item->get_masse,
-				COL_BRAS,  $item->get_bras,
-				COL_BRAS_L,$item->get_bras_l,
-				COL_IMG,   $item->get_img,
-				COL_EST_ITEM, Glib::TRUE,
-				COL_GRAS,   GRAS_NORMAL,
-				COL_PRESENT_PESEE, $item->is_present_pesee,
-				COL_CONFIG_BASE, $item->is_config_base,
-				COL_DRAGABLE, $item->is_dragable,
-				COL_ID, INVALID_ID,
-				COL_EST_TOP, Glib::FALSE);
-			}
-		}
-		else {
-			$treestore->set($mainiter, COL_EST_ITEM, Glib::TRUE, COL_GRAS, GRAS_NORMAL);
-		}
-
-	}
-
-	return $treestore;
 }
 
 
@@ -313,7 +317,9 @@ sub get_categories_profils {
 ###################################"
 sub _drag_received {
 	my ($treeview, $context, $x, $y, $selection, $info, $etime, $args) = @_;
-	my ($drag_iter, $base_items) = @$args;
+#	my ($drag_iter, $base_items) = @$args;
+	my ($ref_drag_iter) = @$args;
+	my $drag_iter = $$ref_drag_iter;
 	my ($path, $position) = $treeview->get_dest_row_at_pos($x, $y);
 	my $treestore = $treeview->get_model;
 	my $niter;
@@ -340,7 +346,7 @@ sub _drag_received {
 
 			if($position eq 'into-or-before') {
 				if ($depth == 2) {
-					if (_is_drop_accepted($treeview, $drag_iter, $base_items)) {
+					if (_is_drop_accepted($treeview, $drag_iter)) {
 						$niter = $treestore->insert_before(undef, $iter);
 					}
 					else {
@@ -353,7 +359,7 @@ sub _drag_received {
 				}
 				else {
 					if (_is_drop_accepted($treeview, $drag_iter)) {
-						_add_iter($treestore, $iter, $drag_iter, $base_items);
+						_add_iter($treestore, $iter, $drag_iter);
 						$niter = $treestore->insert_before($iter, undef);
 					}
 					else {
@@ -364,7 +370,7 @@ sub _drag_received {
 			}
 			else {#if ($position eq 'into-or-after') {
 				if ($depth == 2) {
-					if (_is_drop_accepted($treeview, $drag_iter, $base_items)) {
+					if (_is_drop_accepted($treeview, $drag_iter)) {
 						$niter = $treestore->insert_after(undef, $iter);
 					}
 					else {
@@ -376,7 +382,7 @@ sub _drag_received {
 					$niter = $treestore->insert_after(undef, $iter);
 				}
 				else {
-					if (_is_drop_accepted($treeview, $drag_iter, $base_items)) {
+					if (_is_drop_accepted($treeview, $drag_iter)) {
 						_add_iter($treestore, $iter, $drag_iter);
 						$niter = $treestore->insert_after($iter, undef);
 					}
@@ -436,45 +442,48 @@ sub _drag_received {
 
 #renvoie undef si pas possible
 sub _is_drop_accepted {
-	my ($treeview, $drag_iter, $base_items) = @_;
-	my $id = $treeview->get_model->get($drag_iter, COL_ID);
+#	my ($treeview, $drag_iter) = @_;
+	return _msg_confirm_profil(@_, "Êtes-vous sûr de vouloir le déplacer ?");
 
-	return 1 if ($id == INVALID_ID);
+}
+
+sub _msg_confirm_profil {
+	my ($treeview, $iter, $sConfirm) = @_;
+	my ($id, $str) = $treeview->get_model->get($iter, COL_ID, COL_PROFILS);
+
+	return 1 if ($id == INVALID_ID || !$str);
 
 
-	my $hash = $base_items->[$id];
-	my $liste_profils = $hash->{profils};
+#	my $hash = $base_items->[$id];
+#	my $liste_profils = $hash->{profils};
 	#true
 
-	return 1 unless (scalar(@$liste_profils));
+#	return 1 unless (scalar(@$liste_profils));
 
-	my $str = join ( ', ', map { $_->get_nom } @$liste_profils );
-	#foreach my $profil (@$liste_profils) {
-		#$str .= $profil->get_nom.", ";
-	#}
+#	my $str = join ( ', ', map { $_->get_nom } @$liste_profils );
 	my $dialog = Gtk2::MessageDialog->new (undef, 'modal', 'question', 'yes-no',
-		'Cet optionnel ne sera plus lié aux profils suivants : '.$str.'. Êtes-vous sûr de vouloir le déplacer ?');
+		"Cet optionnel ne sera plus lié aux profils suivants : $str. $sConfirm");
 	my $ret = $dialog->run();
 	$dialog->destroy();
 
 	if ($ret eq 'yes') {
-		delete $hash->{item};
+#		delete $hash->{item};
 		return 1;
 	}
 	else {
 		return 0;
 	}
-
 }
 
 sub _drag_get {
-	my ($treeview, $context, $selection, $target_id, $etime, $drag_iter) = @_;
+	my ($treeview, $context, $selection, $target_id, $etime, $ref_drag_iter) = @_;
 	my ($model, $iter) =  $treeview->get_selection()->get_selected;
-	$drag_iter->set($iter);
+	$$ref_drag_iter = $iter;
 }
 
 sub _drag_delete {
-	my ($treeview, $context, $drag_iter) = @_;
+	my ($treeview, $context, $ref_drag_iter) = @_;
+	my $drag_iter = $$ref_drag_iter;
 	my $treestore = $treeview->get_model;
 	my $dest_iter = $treestore->iter_parent ($drag_iter);
 	if ($dest_iter) {
@@ -487,37 +496,14 @@ sub _drag_delete {
 # Callback de keypress sur le treeview (supprimer une ligne avec suppr)
 #####################################################################
 sub _key_press {
-	my ($treeview, $event, $base_items) = @_;
+	my ($treeview, $event) = @_;
 	return Glib::FALSE unless($event->keyval == $Gtk2::Gdk::Keysyms{Delete});
 
 	my ($model, $iter) =  $treeview->get_selection()->get_selected;
 
 	if ($iter) {
-		my $id = $treeview->get_model->get($iter, COL_ID);
-
-		if ($id == INVALID_ID) {
-			$model->remove($iter);
-			return Glib::TRUE;
-		}
-
-		my $hash = $base_items->[$id];
-		my $liste_profils = $hash->{profils};
-
-		if (scalar(@$liste_profils)) {
-			my $str = join ( ', ', map { $_->get_nom } @$liste_profils );
-			my $dialog = Gtk2::MessageDialog->new (undef, 'modal', 'question', 'yes-no',
-				'Cet optionnel est lié aux profils suivants : '.$str.'. Êtes-vous sûr de vouloir le supprimer ?');
-
-			if ($dialog->run() eq 'yes') {
-				$model->remove($iter);
-				delete $hash->{item};
-			}
-			$dialog->destroy();
-
-		}
-		else {
-			$model->remove($iter);
-		}
+		my ($id, $str) = $treeview->get_model->get($iter, COL_ID, COL_PROFILS);
+		$model->remove($iter) if _msg_confirm_profil($treeview, $iter, "Êtes-vous sûr de vouloir le supprimer ?");
 	}
 	return Glib::TRUE;
 
@@ -533,40 +519,41 @@ sub _key_press {
 
 # Créer une nouvelle colonne toujours éditable
 sub _new_col_editable {
-	my ($treestore, $nom, $col) = @_;
-	my $cellrendrer = _new_renderer($treestore, $col);
+	my ($treeview, $nom, $col) = @_;
+	my $cellrendrer = _new_renderer($treeview, $col);
 	$cellrendrer->set(editable => Glib::TRUE);
 	return (Gtk2::TreeViewColumn->new_with_attributes ($nom, $cellrendrer, 'text', $col, 'weight', COL_GRAS));
 }
 
 # Créer une nouvelle colonne éditable selon la colonne $coleditable
 sub _new_col {
-	my ($treestore, $nom, $col, $coleditable) = @_;
-	my $cellrendrer = _new_renderer($treestore, $col);
+	my ($treeview, $nom, $col, $coleditable) = @_;
+	my $cellrendrer = _new_renderer($treeview, $col);
 	my $tree_column = Gtk2::TreeViewColumn->new_with_attributes ($nom, $cellrendrer,
 		'text', $col, 'editable', $coleditable,	'weight', COL_GRAS);
 	return $tree_column;
 }
 
 sub _new_col_toggle {
-	my ($treestore, $nom, $col) = @_;
-	my $cellrendrer = _new_renderer_toggle($treestore, $col);
+	my ($treeview, $nom, $col) = @_;
+	my $cellrendrer = _new_renderer_toggle($treeview, $col);
 	$cellrendrer->set(activatable => Glib::TRUE);
 	return (Gtk2::TreeViewColumn->new_with_attributes ($nom, $cellrendrer, 'active', $col));
 }
 
 sub _new_col_toggle_may_visible {
-	my ($treestore, $nom, $col, $col_visible) = @_;
-	my $cellrendrer = _new_renderer_toggle($treestore, $col);
+	my ($treeview, $nom, $col, $col_visible) = @_;
+	my $cellrendrer = _new_renderer_toggle($treeview, $col);
 	$cellrendrer->set(activatable => Glib::TRUE);
 	return (Gtk2::TreeViewColumn->new_with_attributes ($nom, $cellrendrer, 'active', $col, 'visible', $col_visible));
 }
 #Créer un nouveau CellRendererText avec le callbaack qui le modifie effectivement en cas d'édition
 sub _new_renderer {
-	my ($treestore, $col) = @_;
+	my ($treeview, $col) = @_;
 	my $cellrendrer = Gtk2::CellRendererText->new;
 	my $func   = sub {
 		my ($cellrenderertext, $path, $new_text) = @_;
+		my $treestore = $treeview->get_model;
 		my $iter = $treestore->get_iter_from_string($path);
 
 		my $dest_iter = $treestore->iter_parent ($iter);
@@ -588,10 +575,11 @@ sub _new_renderer {
 }
 
 sub _new_renderer_toggle {
-	my ($treestore, $col) = @_;
+	my ($treeview, $col) = @_;
 	my $cellrendrer = Gtk2::CellRendererToggle->new;
 	my $func   = sub {
 		my ($cellrenderertext, $path, $new_text) = @_;
+		my $treestore = $treeview->get_model;
 		my $iter = $treestore->get_iter_from_string($path);
 		$treestore->set($iter, $col, not $treestore->get($iter, $col));
 		#print $treestore->get($iter,COL_NOM). ' '.$treestore->get($iter,COL_PRESENT_PESEE)."\n";
